renamed requests modules to be consistent
