+++ views/__init__.py
@@ -1,6 +1,6 @@
 
-from .post_request import get_single_post, get_all_posts
+from .post_requests import get_single_post, get_all_posts
 from .user_requests import get_single_user, get_all_users, get_user_by_username
 from .auth_requests import login_user, create_user
-from .comment_request import get_all_comments_by_post, create_comment, delete_comment, get_all_comments, get_single_comment
+from .comment_requests import get_all_comments_by_post, create_comment, delete_comment, get_all_comments, get_single_comment
